refactored most Player code into Player class
The code for Game.on_update can't be moved into the Player class yet, as the process_movement function needs access to a couple of properties that Player doesn't have direct access to.

I will have to think about how to refactor this next time so the update code can reside in the Player class (or if that is what I want to do at all).