fixed key press issue
If a key is pressed offscreen and released onscreen, or vice versa, the key event handlers don't register, which can result in trying to add a key that's already present, or remove one that isn't.  I added if statements here to make sure we're acting on a state that is still relevant.