begin transition of game screens into Views
I have moved all of the world-specific logic (I'm defining "world" here as whenever the player has control of the main character, so outside of menu windows) into its own View so that it will be easier to switch to other Views, like the inventory and shop window, later.

I imagine the world view will eventually be more complex as well, given that the game will have multiple areas, but that has not been implemented yet.